say anything 1000 times
it does not have to be 1000 times. change the value in the second line to whatever you'd like. 
note change the value before the second semicolon.
to change the thing said, type in between quotations on the third line.
to run you may copy and paste into dartpad.dev
DO NOT REMOVE SEMICOLONS.